requirements.txt
fastapi
uvicorn
qdrant-client
openai
pydantic
python-dotenv
google-genai